+++ Park-Swift/components/ListSpaceButton.js
@@ -3,11 +3,12 @@ import { View, TouchableOpacity, Text, StyleSheet } from "react-native";
 import { useNavigation } from "@react-navigation/native";
 import ThankYouScreen from "../screens/confirmation";
 
-export default function ListSpaceButton() {
+export default function ListSpaceButton({onPress}) {
     const navigation = useNavigation();
 
     const handlePress = () => {
         // Navigate to the "ThankYou" screen
+        onPress();
         navigation.navigate('ThankYou');
     };
     return (
