+++ Park-Swift/screens/ListYourSpaceScreen.js
@@ -1,296 +1,1 @@
-import * as React from 'react';
-import { View, Text, TextInput, TouchableOpacity, StyleSheet, ScrollView } from 'react-native';
-import { Calendar } from 'react-native-calendars';
-import RNPickerSelect from 'react-native-picker-select';
-import { MaterialCommunityIcons } from '@expo/vector-icons';
-import ListSpaceButton from '../components/ListSpaceButton';
- 
-
-
-function ListYourSpaceScreen() {
- const [price, setPrice] = React.useState('');
- const [rentalPeriod, setRentalPeriod] = React.useState('hour');
- const [isNegotiable, setIsNegotiable] = React.useState(null);
- const [selectedDates, setSelectedDates] = React.useState({});
- const [location, setLocation] = React.useState('');
-
-
- return (
-   <View style={styles.container}>
-     <View style={styles.titleRow}>
-       <MaterialCommunityIcons name="arrow-left" size={24} color="black" />
-       <Text style={styles.title}>List Your Space</Text>
-     </View>
-     <ScrollView style={styles.scrollView}>
-     <View>
-     <Text style={styles.headerText}>Location</Text>
-     <View style={styles.inputWithIcon}>
-       <MaterialCommunityIcons name="magnify" size={20} color="black" style={styles.iconInsideInput} />
-       <TextInput
-         style={[styles.inputRounded, styles.inputLocation]}
-         placeholder="Address"
-         value={location}
-         onChangeText={setLocation}
-       />
-     </View>
-
-
-     <Text style={styles.headerText}>Price</Text>
-     <View style={styles.inputWithIcon}>
-       <MaterialCommunityIcons name="currency-usd" size={20} color="black" style={styles.iconInsideInput} />
-       <TextInput
-         style={[styles.inputRounded, styles.inputPrice]}
-         keyboardType="numeric"
-         placeholder="0.00"
-         value={price}
-         onChangeText={setPrice}
-       />
-       <Text style={styles.slash}>/</Text>
-     
-       <RNPickerSelect
-         onValueChange={(value) => setRentalPeriod(value)}
-         items={[
-           { label: 'Hour', value: 'hour' },
-           { label: 'Day', value: 'day' },
-           { label: 'Week', value: 'week' },
-           { label: 'Month', value: 'month' },
-         ]}
-         style={pickerSelectStyles}
-         value={rentalPeriod}
-         useNativeAndroidPickerStyle={false}
-         placeholder={{}}
-         Icon={() => {
-           return <MaterialCommunityIcons name="triangle" size={20} color="lightgrey" style={{ alignSelf: 'center', marginRight: 10, transform: [{ rotate: '180deg' }] }} />;
-         }}
-       />
-
-
-
-
-     </View>
-
-
-     <Text style={styles.headerText}>Negotiable?</Text>
-     <View style={styles.negotiableContainer}>
-       <TouchableOpacity
-         style={styles.option}
-         onPress={() => setIsNegotiable(true)}>
-         <View style={[styles.circle, styles.lightGreyCircle]}>
-           {isNegotiable === true && (
-             <MaterialCommunityIcons name="check" size={24} color="black" />
-           )}
-         </View>
-         <Text style={styles.optionText}>Yes</Text>
-       </TouchableOpacity>
-       <TouchableOpacity
-         style={[styles.option, styles.noOption]}
-         onPress={() => setIsNegotiable(false)}>
-         <View style={[styles.circle, styles.lightGreyCircle]}>
-           {isNegotiable === false && (
-             <MaterialCommunityIcons name="check" size={24} color="black" />
-           )}
-         </View>
-         <Text style={styles.optionText}>No</Text>
-       </TouchableOpacity>
-     </View>
-
-
-     <Text style={styles.centeredHeaderText}>Available Dates</Text>
-     <Calendar
- style={styles.calendarStyle}
- onDayPress={(day) => {
-   // Check if the day is already selected
-   const isSelected = selectedDates[day.dateString] ? true : false;
-
-   setSelectedDates(prevDates => {
-     if (isSelected) {
-       // If selected, remove the selection
-       const updatedDates = { ...prevDates };
-       delete updatedDates[day.dateString]; // Remove the key for the unselected date
-       return updatedDates;
-     } else {
-       // If not selected, add the selection
-       return {
-         ...prevDates,
-         [day.dateString]: { selected: true, selectedColor: 'grey' }
-       };
-     }
-   });
- }}
- markedDates={selectedDates}
-/>
-
-</View>
-</ScrollView>
-
-<ListSpaceButton/>
-
-   </View>
- );
-}
-
-
-const styles = StyleSheet.create({
- container: {
-   top:-120,
-   flex: 1,
-   padding: 10,
-   paddingTop: 150,
-   backgroundColor: '#fff',
-  
- },
- titleRow: {
-   flexDirection: 'row',
-   alignItems: 'center',
-   marginBottom: 20,
- },
- title: {
-   fontSize: 30,
-   color: 'black',
-   fontWeight: 'bold',
-   marginLeft: 10,
- },
- headerText: {
-   fontSize: 20,
-   color: 'black',
-   fontWeight: 'bold',
-   marginTop: 15,
- },
- inputWithIcon: {
-   flexDirection: 'row',
-   alignItems: 'center',
-   marginTop: 20,
-  
- },
- iconInsideInput: {
-   position: 'absolute',
-   marginLeft: 15,
-   zIndex: 1,
- },
- inputRounded: {
-   borderWidth: 1,
-   borderRadius: 20,
-   padding: 10,
-   backgroundColor: '#f0f0f0',
-   fontSize: 16,
-   flex: 1,
-   height: 50,
-   paddingLeft: 45,
-   borderColor: 'transparent',
- },
- inputPrice: {
-   marginRight: 10,
-   flex: 0,
-   width: 160,
-   height: 45,
-   marginTop: 6,
-   paddingLeft: 45,
- },
- slash: {
-   fontSize: 40,
-   color: 'black',
-   marginHorizontal: 5,
- },
- negotiableContainer: {
-   flexDirection: 'row',
-   marginTop: 15,
- },
- option: {
-   flexDirection: 'row',
-   alignItems: 'center',
-   marginRight: 30,
- },
- noOption: {
-   marginLeft: 10,
- },
- circle: {
-   width: 24,
-   height: 24,
-   borderRadius: 12,
-   borderWidth: 1,
-   borderColor: 'transparent',
-   justifyContent: 'center',
-   alignItems: 'center',
-   backgroundColor: '#f0f0f0',
- },
- optionText: {
-   fontSize: 16,
-   color: 'black',
-   marginLeft: 5,
- },
- centeredHeaderText: {
-   fontSize: 20,
-   color: 'black',
-   fontWeight: 'bold',
-   marginTop: 15,
-   textAlign: 'center',
- },
- calendarStyle: {
- marginTop: 10,
- color: 'black',
- borderWidth: 1,
- borderColor: '#d3d3d3',
- justifyContent: 'center',
- },
-});
-
-
-const pickerSelectStyles = StyleSheet.create({
- inputIOS: {
-   fontSize: 16,
-   paddingVertical: 10,
-   paddingHorizontal: 10,
-   borderWidth: 1,
-   borderColor: 'transparent',
-   borderRadius: 15,
-   color: 'black',
-   paddingRight: 30,
-   backgroundColor: '#f0f0f0',
-   height: 45,
-   marginTop: 6,
-   width: 125,
- 
-  
- },
- inputAndroid: {
-   fontSize: 16,
-   paddingHorizontal: 10,
-   paddingVertical: 8,
-   borderWidth: 0.5,
-   borderColor: 'transparent',
-   borderRadius: 8,
-   color: 'black',
-   paddingRight: 30,
-   backgroundColor: '#f0f0f0',
-   height: 45,
-   marginTop: -5,
-   width: 125,
- },
- iconContainer: {
-   top: '50%',
-   right: 0,
-   transform: [{ translateY: -10 }], // Adjust this value as needed to center the icon
- },
- button: {
-  alignItems: 'center',
-  justifyContent: 'center',
-  paddingVertical: 12,
-  paddingHorizontal: 32,
-  borderRadius: 4,
-  elevation: 3,
-  backgroundColor: 'black',
-},
-text: {
-  fontSize: 16,
-  lineHeight: 21,
-  fontWeight: 'bold',
-  letterSpacing: 0.25,
-  color: 'white',
-},
-});
-
-
-export default ListYourSpaceScreen;
-
-
-
+105,109,112,111,114,116,32,42,32,97,115,32,82,101,97,99,116,32,102,114,111,109,32,39,114,101,97,99,116,39,59,10,105,109,112,111,114,116,32,123,32,86,105,101,119,44,32,84,101,120,116,44,32,84,101,120,116,73,110,112,117,116,44,32,84,111,117,99,104,97,98,108,101,79,112,97,99,105,116,121,44,32,83,116,121,108,101,83,104,101,101,116,44,32,83,99,114,111,108,108,86,105,101,119,32,125,32,102,114,111,109,32,39,114,101,97,99,116,45,110,97,116,105,118,101,39,59,10,105,109,112,111,114,116,32,123,32,67,97,108,101,110,100,97,114,32,125,32,102,114,111,109,32,39,114,101,97,99,116,45,110,97,116,105,118,101,45,99,97,108,101,110,100,97,114,115,39,59,10,105,109,112,111,114,116,32,82,78,80,105,99,107,101,114,83,101,108,101,99,116,32,102,114,111,109,32,39,114,101,97,99,116,45,110,97,116,105,118,101,45,112,105,99,107,101,114,45,115,101,108,101,99,116,39,59,10,105,109,112,111,114,116,32,123,32,77,97,116,101,114,105,97,108,67,111,109,109,117,110,105,116,121,73,99,111,110,115,32,125,32,102,114,111,109,32,39,64,101,120,112,111,47,118,101,99,116,111,114,45,105,99,111,110,115,39,59,10,105,109,112,111,114,116,32,76,105,115,116,83,112,97,99,101,66,117,116,116,111,110,32,102,114,111,109,32,39,46,46,47,99,111,109,112,111,110,101,110,116,115,47,76,105,115,116,83,112,97,99,101,66,117,116,116,111,110,39,59,10,105,109,112,111,114,116,32,123,32,99,114,101,97,116,101,80,111,115,116,32,125,32,102,114,111,109,32,39,46,46,47,102,105,114,101,98,97,115,101,70,117,110,99,116,105,111,110,115,47,102,105,114,101,98,97,115,101,39,59,10,10,32,10,10,10,102,117,110,99,116,105,111,110,32,67,114,101,97,116,101,80,111,115,116,40,123,32,110,97,118,105,103,97,116,105,111,110,44,32,114,111,117,116,101,32,125,41,32,123,10,32,32,99,111,110,115,116,32,117,115,101,114,73,100,32,61,32,114,111,117,116,101,46,112,97,114,97,109,115,46,117,115,101,114,73,100,59,10,32,32,99,111,110,115,116,32,91,108,111,99,97,116,105,111,110,44,32,115,101,116,76,111,99,97,116,105,111,110,93,32,61,32,82,101,97,99,116,46,117,115,101,83,116,97,116,101,40,39,39,41,59,10,32,32,99,111,110,115,116,32,91,112,114,105,99,101,44,32,115,101,116,80,114,105,99,101,93,32,61,32,82,101,97,99,116,46,117,115,101,83,116,97,116,101,40,39,39,41,59,10,32,32,99,111,110,115,116,32,91,114,101,110,116,97,108,80,101,114,105,111,100,44,32,115,101,116,82,101,110,116,97,108,80,101,114,105,111,100,93,32,61,32,82,101,97,99,116,46,117,115,101,83,116,97,116,101,40,39,104,111,117,114,39,41,59,10,32,32,99,111,110,115,116,32,91,105,115,78,101,103,111,116,105,97,98,108,101,44,32,115,101,116,73,115,78,101,103,111,116,105,97,98,108,101,93,32,61,32,82,101,97,99,116,46,117,115,101,83,116,97,116,101,40,110,117,108,108,41,59,10,32,32,99,111,110,115,116,32,91,115,101,108,101,99,116,101,100,68,97,116,101,115,44,32,115,101,116,83,101,108,101,99,116,101,100,68,97,116,101,115,93,32,61,32,82,101,97,99,116,46,117,115,101,83,116,97,116,101,40,123,125,41,59,10,32,10,32,99,111,110,115,116,32,111,110,80,111,115,116,80,114,101,115,115,32,61,32,40,41,32,61,62,32,123,10,32,32,32,97,108,101,114,116,40,117,115,101,114,73,100,41,59,10,32,32,32,99,114,101,97,116,101,80,111,115,116,40,117,115,101,114,73,100,44,32,108,111,99,97,116,105,111,110,44,32,114,101,110,116,97,108,80,101,114,105,111,100,44,32,112,114,105,99,101,44,32,105,115,78,101,103,111,116,105,97,98,108,101,44,32,115,101,108,101,99,116,101,100,68,97,116,101,115,41,10,32,32,32,32,32,46,116,104,101,110,40,40,41,32,61,62,32,123,10,32,32,32,32,32,32,32,110,97,118,105,103,97,116,105,111,110,46,110,97,118,105,103,97,116,101,40,39,72,111,109,101,39,41,59,10,32,32,32,32,32,125,41,10,32,32,32,32,32,46,99,97,116,99,104,40,40,101,114,114,111,114,41,32,61,62,32,123,10,32,32,32,32,32,32,32,99,111,110,115,111,108,101,46,101,114,114,111,114,40,39,69,114,114,111,114,32,99,114,101,97,116,105,110,103,32,112,111,115,116,58,39,44,32,101,114,114,111,114,41,59,10,32,32,32,32,32,32,32,47,47,32,72,97,110,100,108,101,32,116,104,101,32,101,114,114,111,114,32,104,101,114,101,44,32,101,46,103,46,44,32,115,104,111,119,32,97,110,32,101,114,114,111,114,32,109,101,115,115,97,103,101,32,116,111,32,116,104,101,32,117,115,101,114,10,32,32,32,32,32,125,41,59,10,32,125,10,32,10,32,114,101,116,117,114,110,32,40,10,32,32,32,60,86,105,101,119,32,115,116,121,108,101,61,123,115,116,121,108,101,115,46,99,111,110,116,97,105,110,101,114,125,62,10,32,32,32,32,32,60,86,105,101,119,32,115,116,121,108,101,61,123,115,116,121,108,101,115,46,116,105,116,108,101,82,111,119,125,62,10,32,32,32,32,32,32,32,60,77,97,116,101,114,105,97,108,67,111,109,109,117,110,105,116,121,73,99,111,110,115,32,110,97,109,101,61,34,97,114,114,111,119,45,108,101,102,116,34,32,115,105,122,101,61,123,50,52,125,32,99,111,108,111,114,61,34,98,108,97,99,107,34,32,47,62,10,32,32,32,32,32,32,32,60,84,101,120,116,32,115,116,121,108,101,61,123,115,116,121,108,101,115,46,116,105,116,108,101,125,62,76,105,115,116,32,89,111,117,114,32,83,112,97,99,101,60,47,84,101,120,116,62,10,32,32,32,32,32,60,47,86,105,101,119,62,10,32,32,32,32,32,60,83,99,114,111,108,108,86,105,101,119,32,115,116,121,108,101,61,123,115,116,121,108,101,115,46,115,99,114,111,108,108,86,105,101,119,125,62,10,32,32,32,32,32,60,86,105,101,119,62,10,32,32,32,32,32,60,84,101,120,116,32,115,116,121,108,101,61,123,115,116,121,108,101,115,46,104,101,97,100,101,114,84,101,120,116,125,62,76,111,99,97,116,105,111,110,60,47,84,101,120,116,62,10,32,32,32,32,32,60,86,105,101,119,32,115,116,121,108,101,61,123,115,116,121,108,101,115,46,105,110,112,117,116,87,105,116,104,73,99,111,110,125,62,10,32,32,32,32,32,32,32,60,77,97,116,101,114,105,97,108,67,111,109,109,117,110,105,116,121,73,99,111,110,115,32,110,97,109,101,61,34,109,97,103,110,105,102,121,34,32,115,105,122,101,61,123,50,48,125,32,99,111,108,111,114,61,34,98,108,97,99,107,34,32,115,116,121,108,101,61,123,115,116,121,108,101,115,46,105,99,111,110,73,110,115,105,100,101,73,110,112,117,116,125,32,47,62,10,32,32,32,32,32,32,32,60,84,101,120,116,73,110,112,117,116,10,32,32,32,32,32,32,32,32,32,115,116,121,108,101,61,123,91,115,116,121,108,101,115,46,105,110,112,117,116,82,111,117,110,100,101,100,44,32,115,116,121,108,101,115,46,105,110,112,117,116,76,111,99,97,116,105,111,110,93,125,10,32,32,32,32,32,32,32,32,32,112,108,97,99,101,104,111,108,100,101,114,61,34,65,100,100,114,101,115,115,34,10,32,32,32,32,32,32,32,32,32,118,97,108,117,101,61,123,108,111,99,97,116,105,111,110,125,10,32,32,32,32,32,32,32,32,32,111,110,67,104,97,110,103,101,84,101,120,116,61,123,115,101,116,76,111,99,97,116,105,111,110,125,10,32,32,32,32,32,32,32,47,62,10,32,32,32,32,32,60,47,86,105,101,119,62,10,10,10,32,32,32,32,32,60,84,101,120,116,32,115,116,121,108,101,61,123,115,116,121,108,101,115,46,104,101,97,100,101,114,84,101,120,116,125,62,80,114,105,99,101,60,47,84,101,120,116,62,10,32,32,32,32,32,60,86,105,101,119,32,115,116,121,108,101,61,123,115,116,121,108,101,115,46,105,110,112,117,116,87,105,116,104,73,99,111,110,125,62,10,32,32,32,32,32,32,32,60,77,97,116,101,114,105,97,108,67,111,109,109,117,110,105,116,121,73,99,111,110,115,32,110,97,109,101,61,34,99,117,114,114,101,110,99,121,45,117,115,100,34,32,115,105,122,101,61,123,50,48,125,32,99,111,108,111,114,61,34,98,108,97,99,107,34,32,115,116,121,108,101,61,123,115,116,121,108,101,115,46,105,99,111,110,73,110,115,105,100,101,73,110,112,117,116,125,32,47,62,10,32,32,32,32,32,32,32,60,84,101,120,116,73,110,112,117,116,10,32,32,32,32,32,32,32,32,32,115,116,121,108,101,61,123,91,115,116,121,108,101,115,46,105,110,112,117,116,82,111,117,110,100,101,100,44,32,115,116,121,108,101,115,46,105,110,112,117,116,80,114,105,99,101,93,125,10,32,32,32,32,32,32,32,32,32,107,101,121,98,111,97,114,100,84,121,112,101,61,34,110,117,109,101,114,105,99,34,10,32,32,32,32,32,32,32,32,32,112,108,97,99,101,104,111,108,100,101,114,61,34,48,46,48,48,34,10,32,32,32,32,32,32,32,32,32,118,97,108,117,101,61,123,112,114,105,99,101,125,10,32,32,32,32,32,32,32,32,32,111,110,67,104,97,110,103,101,84,101,120,116,61,123,115,101,116,80,114,105,99,101,125,10,32,32,32,32,32,32,32,47,62,10,32,32,32,32,32,32,32,60,84,101,120,116,32,115,116,121,108,101,61,123,115,116,121,108,101,115,46,115,108,97,115,104,125,62,47,60,47,84,101,120,116,62,10,32,32,32,32,32,10,32,32,32,32,32,32,32,60,82,78,80,105,99,107,101,114,83,101,108,101,99,116,10,32,32,32,32,32,32,32,32,32,111,110,86,97,108,117,101,67,104,97,110,103,101,61,123,40,118,97,108,117,101,41,32,61,62,32,115,101,116,82,101,110,116,97,108,80,101,114,105,111,100,40,118,97,108,117,101,41,125,10,32,32,32,32,32,32,32,32,32,105,116,101,109,115,61,123,91,10,32,32,32,32,32,32,32,32,32,32,32,123,32,108,97,98,101,108,58,32,39,72,111,117,114,39,44,32,118,97,108,117,101,58,32,39,104,111,117,114,39,32,125,44,10,32,32,32,32,32,32,32,32,32,32,32,123,32,108,97,98,101,108,58,32,39,68,97,121,39,44,32,118,97,108,117,101,58,32,39,100,97,121,39,32,125,44,10,32,32,32,32,32,32,32,32,32,32,32,123,32,108,97,98,101,108,58,32,39,87,101,101,107,39,44,32,118,97,108,117,101,58,32,39,119,101,101,107,39,32,125,44,10,32,32,32,32,32,32,32,32,32,32,32,123,32,108,97,98,101,108,58,32,39,77,111,110,116,104,39,44,32,118,97,108,117,101,58,32,39,109,111,110,116,104,39,32,125,44,10,32,32,32,32,32,32,32,32,32,93,125,10,32,32,32,32,32,32,32,32,32,115,116,121,108,101,61,123,112,105,99,107,101,114,83,101,108,101,99,116,83,116,121,108,101,115,125,10,32,32,32,32,32,32,32,32,32,118,97,108,117,101,61,123,114,101,110,116,97,108,80,101,114,105,111,100,125,10,32,32,32,32,32,32,32,32,32,117,115,101,78,97,116,105,118,101,65,110,100,114,111,105,100,80,105,99,107,101,114,83,116,121,108,101,61,123,102,97,108,115,101,125,10,32,32,32,32,32,32,32,32,32,112,108,97,99,101,104,111,108,100,101,114,61,123,123,125,125,10,32,32,32,32,32,32,32,32,32,73,99,111,110,61,123,40,41,32,61,62,32,123,10,32,32,32,32,32,32,32,32,32,32,32,114,101,116,117,114,110,32,60,77,97,116,101,114,105,97,108,67,111,109,109,117,110,105,116,121,73,99,111,110,115,32,110,97,109,101,61,34,116,114,105,97,110,103,108,101,34,32,115,105,122,101,61,123,50,48,125,32,99,111,108,111,114,61,34,108,105,103,104,116,103,114,101,121,34,32,115,116,121,108,101,61,123,123,32,97,108,105,103,110,83,101,108,102,58,32,39,99,101,110,116,101,114,39,44,32,109,97,114,103,105,110,82,105,103,104,116,58,32,49,48,44,32,116,114,97,110,115,102,111,114,109,58,32,91,123,32,114,111,116,97,116,101,58,32,39,49,56,48,100,101,103,39,32,125,93,32,125,125,32,47,62,59,10,32,32,32,32,32,32,32,32,32,125,125,10,32,32,32,32,32,32,32,47,62,10,10,10,10,10,32,32,32,32,32,60,47,86,105,101,119,62,10,10,10,32,32,32,32,32,60,84,101,120,116,32,115,116,121,108,101,61,123,115,116,121,108,101,115,46,104,101,97,100,101,114,84,101,120,116,125,62,78,101,103,111,116,105,97,98,108,101,63,60,47,84,101,120,116,62,10,32,32,32,32,32,60,86,105,101,119,32,115,116,121,108,101,61,123,115,116,121,108,101,115,46,110,101,103,111,116,105,97,98,108,101,67,111,110,116,97,105,110,101,114,125,62,10,32,32,32,32,32,32,32,60,84,111,117,99,104,97,98,108,101,79,112,97,99,105,116,121,10,32,32,32,32,32,32,32,32,32,115,116,121,108,101,61,123,115,116,121,108,101,115,46,111,112,116,105,111,110,125,10,32,32,32,32,32,32,32,32,32,111,110,80,114,101,115,115,61,123,40,41,32,61,62,32,115,101,116,73,115,78,101,103,111,116,105,97,98,108,101,40,116,114,117,101,41,125,62,10,32,32,32,32,32,32,32,32,32,60,86,105,101,119,32,115,116,121,108,101,61,123,91,115,116,121,108,101,115,46,99,105,114,99,108,101,44,32,115,116,121,108,101,115,46,108,105,103,104,116,71,114,101,121,67,105,114,99,108,101,93,125,62,10,32,32,32,32,32,32,32,32,32,32,32,123,105,115,78,101,103,111,116,105,97,98,108,101,32,61,61,61,32,116,114,117,101,32,38,38,32,40,10,32,32,32,32,32,32,32,32,32,32,32,32,32,60,77,97,116,101,114,105,97,108,67,111,109,109,117,110,105,116,121,73,99,111,110,115,32,110,97,109,101,61,34,99,104,101,99,107,34,32,115,105,122,101,61,123,50,52,125,32,99,111,108,111,114,61,34,98,108,97,99,107,34,32,47,62,10,32,32,32,32,32,32,32,32,32,32,32,41,125,10,32,32,32,32,32,32,32,32,32,60,47,86,105,101,119,62,10,32,32,32,32,32,32,32,32,32,60,84,101,120,116,32,115,116,121,108,101,61,123,115,116,121,108,101,115,46,111,112,116,105,111,110,84,101,120,116,125,62,89,101,115,60,47,84,101,120,116,62,10,32,32,32,32,32,32,32,60,47,84,111,117,99,104,97,98,108,101,79,112,97,99,105,116,121,62,10,32,32,32,32,32,32,32,60,84,111,117,99,104,97,98,108,101,79,112,97,99,105,116,121,10,32,32,32,32,32,32,32,32,32,115,116,121,108,101,61,123,91,115,116,121,108,101,115,46,111,112,116,105,111,110,44,32,115,116,121,108,101,115,46,110,111,79,112,116,105,111,110,93,125,10,32,32,32,32,32,32,32,32,32,111,110,80,114,101,115,115,61,123,40,41,32,61,62,32,115,101,116,73,115,78,101,103,111,116,105,97,98,108,101,40,102,97,108,115,101,41,125,62,10,32,32,32,32,32,32,32,32,32,60,86,105,101,119,32,115,116,121,108,101,61,123,91,115,116,121,108,101,115,46,99,105,114,99,108,101,44,32,115,116,121,108,101,115,46,108,105,103,104,116,71,114,101,121,67,105,114,99,108,101,93,125,62,10,32,32,32,32,32,32,32,32,32,32,32,123,105,115,78,101,103,111,116,105,97,98,108,101,32,61,61,61,32,102,97,108,115,101,32,38,38,32,40,10,32,32,32,32,32,32,32,32,32,32,32,32,32,60,77,97,116,101,114,105,97,108,67,111,109,109,117,110,105,116,121,73,99,111,110,115,32,110,97,109,101,61,34,99,104,101,99,107,34,32,115,105,122,101,61,123,50,52,125,32,99,111,108,111,114,61,34,98,108,97,99,107,34,32,47,62,10,32,32,32,32,32,32,32,32,32,32,32,41,125,10,32,32,32,32,32,32,32,32,32,60,47,86,105,101,119,62,10,32,32,32,32,32,32,32,32,32,60,84,101,120,116,32,115,116,121,108,101,61,123,115,116,121,108,101,115,46,111,112,116,105,111,110,84,101,120,116,125,62,78,111,60,47,84,101,120,116,62,10,32,32,32,32,32,32,32,60,47,84,111,117,99,104,97,98,108,101,79,112,97,99,105,116,121,62,10,32,32,32,32,32,60,47,86,105,101,119,62,10,10,10,32,32,32,32,32,60,84,101,120,116,32,115,116,121,108,101,61,123,115,116,121,108,101,115,46,99,101,110,116,101,114,101,100,72,101,97,100,101,114,84,101,120,116,125,62,65,118,97,105,108,97,98,108,101,32,68,97,116,101,115,60,47,84,101,120,116,62,10,32,32,32,32,32,60,67,97,108,101,110,100,97,114,10,32,115,116,121,108,101,61,123,115,116,121,108,101,115,46,99,97,108,101,110,100,97,114,83,116,121,108,101,125,10,32,111,110,68,97,121,80,114,101,115,115,61,123,40,100,97,121,41,32,61,62,32,123,10,32,32,32,47,47,32,67,104,101,99,107,32,105,102,32,116,104,101,32,100,97,121,32,105,115,32,97,108,114,101,97,100,121,32,115,101,108,101,99,116,101,100,10,32,32,32,99,111,110,115,116,32,105,115,83,101,108,101,99,116,101,100,32,61,32,115,101,108,101,99,116,101,100,68,97,116,101,115,91,100,97,121,46,100,97,116,101,83,116,114,105,110,103,93,32,63,32,116,114,117,101,32,58,32,102,97,108,115,101,59,10,10,32,32,32,115,101,116,83,101,108,101,99,116,101,100,68,97,116,101,115,40,112,114,101,118,68,97,116,101,115,32,61,62,32,123,10,32,32,32,32,32,105,102,32,40,105,115,83,101,108,101,99,116,101,100,41,32,123,10,32,32,32,32,32,32,32,47,47,32,73,102,32,115,101,108,101,99,116,101,100,44,32,114,101,109,111,118,101,32,116,104,101,32,115,101,108,101,99,116,105,111,110,10,32,32,32,32,32,32,32,99,111,110,115,116,32,117,112,100,97,116,101,100,68,97,116,101,115,32,61,32,123,32,46,46,46,112,114,101,118,68,97,116,101,115,32,125,59,10,32,32,32,32,32,32,32,100,101,108,101,116,101,32,117,112,100,97,116,101,100,68,97,116,101,115,91,100,97,121,46,100,97,116,101,83,116,114,105,110,103,93,59,32,47,47,32,82,101,109,111,118,101,32,116,104,101,32,107,101,121,32,102,111,114,32,116,104,101,32,117,110,115,101,108,101,99,116,101,100,32,100,97,116,101,10,32,32,32,32,32,32,32,114,101,116,117,114,110,32,117,112,100,97,116,101,100,68,97,116,101,115,59,10,32,32,32,32,32,125,32,101,108,115,101,32,123,10,32,32,32,32,32,32,32,47,47,32,73,102,32,110,111,116,32,115,101,108,101,99,116,101,100,44,32,97,100,100,32,116,104,101,32,115,101,108,101,99,116,105,111,110,10,32,32,32,32,32,32,32,114,101,116,117,114,110,32,123,10,32,32,32,32,32,32,32,32,32,46,46,46,112,114,101,118,68,97,116,101,115,44,10,32,32,32,32,32,32,32,32,32,91,100,97,121,46,100,97,116,101,83,116,114,105,110,103,93,58,32,123,32,115,101,108,101,99,116,101,100,58,32,116,114,117,101,44,32,115,101,108,101,99,116,101,100,67,111,108,111,114,58,32,39,103,114,101,121,39,32,125,10,32,32,32,32,32,32,32,125,59,10,32,32,32,32,32,125,10,32,32,32,125,41,59,10,32,125,125,10,32,109,97,114,107,101,100,68,97,116,101,115,61,123,115,101,108,101,99,116,101,100,68,97,116,101,115,125,10,47,62,10,10,60,47,86,105,101,119,62,10,60,47,83,99,114,111,108,108,86,105,101,119,62,10,10,60,76,105,115,116,83,112,97,99,101,66,117,116,116,111,110,32,111,110,80,114,101,115,115,61,123,40,41,32,61,62,32,99,111,110,115,111,108,101,46,108,111,103,40,39,66,117,116,116,111,110,32,112,114,101,115,115,101,100,33,39,41,125,32,47,62,10,10,10,99,111,110,115,116,32,115,116,121,108,101,115,32,61,32,83,116,121,108,101,83,104,101,101,116,46,99,114,101,97,116,101,40,123,10,32,99,111,110,116,97,105,110,101,114,58,32,123,10,32,32,32,116,111,112,58,45,49,50,48,44,10,32,32,32,102,108,101,120,58,32,49,44,10,32,32,32,112,97,100,100,105,110,103,58,32,49,48,44,10,32,32,32,112,97,100,100,105,110,103,84,111,112,58,32,49,53,48,44,10,32,32,32,98,97,99,107,103,114,111,117,110,100,67,111,108,111,114,58,32,39,35,102,102,102,39,44,10,32,32,10,32,125,44,10,32,116,105,116,108,101,82,111,119,58,32,123,10,32,32,32,102,108,101,120,68,105,114,101,99,116,105,111,110,58,32,39,114,111,119,39,44,10,32,32,32,97,108,105,103,110,73,116,101,109,115,58,32,39,99,101,110,116,101,114,39,44,10,32,32,32,109,97,114,103,105,110,66,111,116,116,111,109,58,32,50,48,44,10,32,125,44,10,32,116,105,116,108,101,58,32,123,10,32,32,32,102,111,110,116,83,105,122,101,58,32,51,48,44,10,32,32,32,99,111,108,111,114,58,32,39,98,108,97,99,107,39,44,10,32,32,32,102,111,110,116,87,101,105,103,104,116,58,32,39,98,111,108,100,39,44,10,32,32,32,109,97,114,103,105,110,76,101,102,116,58,32,49,48,44,10,32,125,44,10,32,104,101,97,100,101,114,84,101,120,116,58,32,123,10,32,32,32,102,111,110,116,83,105,122,101,58,32,50,48,44,10,32,32,32,99,111,108,111,114,58,32,39,98,108,97,99,107,39,44,10,32,32,32,102,111,110,116,87,101,105,103,104,116,58,32,39,98,111,108,100,39,44,10,32,32,32,109,97,114,103,105,110,84,111,112,58,32,49,53,44,10,32,125,44,10,32,105,110,112,117,116,87,105,116,104,73,99,111,110,58,32,123,10,32,32,32,102,108,101,120,68,105,114,101,99,116,105,111,110,58,32,39,114,111,119,39,44,10,32,32,32,97,108,105,103,110,73,116,101,109,115,58,32,39,99,101,110,116,101,114,39,44,10,32,32,32,109,97,114,103,105,110,84,111,112,58,32,50,48,44,10,32,32,10,32,125,44,10,32,105,99,111,110,73,110,115,105,100,101,73,110,112,117,116,58,32,123,10,32,32,32,112,111,115,105,116,105,111,110,58,32,39,97,98,115,111,108,117,116,101,39,44,10,32,32,32,109,97,114,103,105,110,76,101,102,116,58,32,49,53,44,10,32,32,32,122,73,110,100,101,120,58,32,49,44,10,32,125,44,10,32,105,110,112,117,116,82,111,117,110,100,101,100,58,32,123,10,32,32,32,98,111,114,100,101,114,87,105,100,116,104,58,32,49,44,10,32,32,32,98,111,114,100,101,114,82,97,100,105,117,115,58,32,50,48,44,10,32,32,32,112,97,100,100,105,110,103,58,32,49,48,44,10,32,32,32,98,97,99,107,103,114,111,117,110,100,67,111,108,111,114,58,32,39,35,102,48,102,48,102,48,39,44,10,32,32,32,102,111,110,116,83,105,122,101,58,32,49,54,44,10,32,32,32,102,108,101,120,58,32,49,44,10,32,32,32,104,101,105,103,104,116,58,32,53,48,44,10,32,32,32,112,97,100,100,105,110,103,76,101,102,116,58,32,52,53,44,10,32,32,32,98,111,114,100,101,114,67,111,108,111,114,58,32,39,116,114,97,110,115,112,97,114,101,110,116,39,44,10,32,125,44,10,32,105,110,112,117,116,80,114,105,99,101,58,32,123,10,32,32,32,109,97,114,103,105,110,82,105,103,104,116,58,32,49,48,44,10,32,32,32,102,108,101,120,58,32,48,44,10,32,32,32,119,105,100,116,104,58,32,49,54,48,44,10,32,32,32,104,101,105,103,104,116,58,32,52,53,44,10,32,32,32,109,97,114,103,105,110,84,111,112,58,32,54,44,10,32,32,32,112,97,100,100,105,110,103,76,101,102,116,58,32,52,53,44,10,32,125,44,10,32,115,108,97,115,104,58,32,123,10,32,32,32,102,111,110,116,83,105,122,101,58,32,52,48,44,10,32,32,32,99,111,108,111,114,58,32,39,98,108,97,99,107,39,44,10,32,32,32,109,97,114,103,105,110,72,111,114,105,122,111,110,116,97,108,58,32,53,44,10,32,125,44,10,32,110,101,103,111,116,105,97,98,108,101,67,111,110,116,97,105,110,101,114,58,32,123,10,32,32,32,102,108,101,120,68,105,114,101,99,116,105,111,110,58,32,39,114,111,119,39,44,10,32,32,32,109,97,114,103,105,110,84,111,112,58,32,49,53,44,10,32,125,44,10,32,111,112,116,105,111,110,58,32,123,10,32,32,32,102,108,101,120,68,105,114,101,99,116,105,111,110,58,32,39,114,111,119,39,44,10,32,32,32,97,108,105,103,110,73,116,101,109,115,58,32,39,99,101,110,116,101,114,39,44,10,32,32,32,109,97,114,103,105,110,82,105,103,104,116,58,32,51,48,44,10,32,125,44,10,32,110,111,79,112,116,105,111,110,58,32,123,10,32,32,32,109,97,114,103,105,110,76,101,102,116,58,32,49,48,44,10,32,125,44,10,32,99,105,114,99,108,101,58,32,123,10,32,32,32,119,105,100,116,104,58,32,50,52,44,10,32,32,32,104,101,105,103,104,116,58,32,50,52,44,10,32,32,32,98,111,114,100,101,114,82,97,100,105,117,115,58,32,49,50,44,10,32,32,32,98,111,114,100,101,114,87,105,100,116,104,58,32,49,44,10,32,32,32,98,111,114,100,101,114,67,111,108,111,114,58,32,39,116,114,97,110,115,112,97,114,101,110,116,39,44,10,32,32,32,106,117,115,116,105,102,121,67,111,110,116,101,110,116,58,32,39,99,101,110,116,101,114,39,44,10,32,32,32,97,108,105,103,110,73,116,101,109,115,58,32,39,99,101,110,116,101,114,39,44,10,32,32,32,98,97,99,107,103,114,111,117,110,100,67,111,108,111,114,58,32,39,35,102,48,102,48,102,48,39,44,10,32,125,44,10,32,111,112,116,105,111,110,84,101,120,116,58,32,123,10,32,32,32,102,111,110,116,83,105,122,101,58,32,49,54,44,10,32,32,32,99,111,108,111,114,58,32,39,98,108,97,99,107,39,44,10,32,32,32,109,97,114,103,105,110,76,101,102,116,58,32,53,44,10,32,125,44,10,32,99,101,110,116,101,114,101,100,72,101,97,100,101,114,84,101,120,116,58,32,123,10,32,32,32,102,111,110,116,83,105,122,101,58,32,50,48,44,10,32,32,32,99,111,108,111,114,58,32,39,98,108,97,99,107,39,44,10,32,32,32,102,111,110,116,87,101,105,103,104,116,58,32,39,98,111,108,100,39,44,10,32,32,32,109,97,114,103,105,110,84,111,112,58,32,49,53,44,10,32,32,32,116,101,120,116,65,108,105,103,110,58,32,39,99,101,110,116,101,114,39,44,10,32,125,44,10,32,99,97,108,101,110,100,97,114,83,116,121,108,101,58,32,123,10,32,109,97,114,103,105,110,84,111,112,58,32,49,48,44,10,32,99,111,108,111,114,58,32,39,98,108,97,99,107,39,44,10,32,98,111,114,100,101,114,87,105,100,116,104,58,32,49,44,10,32,98,111,114,100,101,114,67,111,108,111,114,58,32,39,35,100,51,100,51,100,51,39,44,10,32,106,117,115,116,105,102,121,67,111,110,116,101,110,116,58,32,39,99,101,110,116,101,114,39,44,10,32,125,44,10,125,41,59,10,10,10,99,111,110,115,116,32,112,105,99,107,101,114,83,101,108,101,99,116,83,116,121,108,101,115,32,61,32,83,116,121,108,101,83,104,101,101,116,46,99,114,101,97,116,101,40,123,10,32,105,110,112,117,116,73,79,83,58,32,123,10,32,32,32,102,111,110,116,83,105,122,101,58,32,49,54,44,10,32,32,32,112,97,100,100,105,110,103,86,101,114,116,105,99,97,108,58,32,49,48,44,10,32,32,32,112,97,100,100,105,110,103,72,111,114,105,122,111,110,116,97,108,58,32,49,48,44,10,32,32,32,98,111,114,100,101,114,87,105,100,116,104,58,32,49,44,10,32,32,32,98,111,114,100,101,114,67,111,108,111,114,58,32,39,116,114,97,110,115,112,97,114,101,110,116,39,44,10,32,32,32,98,111,114,100,101,114,82,97,100,105,117,115,58,32,49,53,44,10,32,32,32,99,111,108,111,114,58,32,39,98,108,97,99,107,39,44,10,32,32,32,112,97,100,100,105,110,103,82,105,103,104,116,58,32,51,48,44,10,32,32,32,98,97,99,107,103,114,111,117,110,100,67,111,108,111,114,58,32,39,35,102,48,102,48,102,48,39,44,10,32,32,32,104,101,105,103,104,116,58,32,52,53,44,10,32,32,32,109,97,114,103,105,110,84,111,112,58,32,54,44,10,32,32,32,119,105,100,116,104,58,32,49,50,53,44,10,32,10,32,32,10,32,125,44,10,32,105,110,112,117,116,65,110,100,114,111,105,100,58,32,123,10,32,32,32,102,111,110,116,83,105,122,101,58,32,49,54,44,10,32,32,32,112,97,100,100,105,110,103,72,111,114,105,122,111,110,116,97,108,58,32,49,48,44,10,32,32,32,112,97,100,100,105,110,103,86,101,114,116,105,99,97,108,58,32,56,44,10,32,32,32,98,111,114,100,101,114,87,105,100,116,104,58,32,48,46,53,44,10,32,32,32,98,111,114,100,101,114,67,111,108,111,114,58,32,39,116,114,97,110,115,112,97,114,101,110,116,39,44,10,32,32,32,98,111,114,100,101,114,82,97,100,105,117,115,58,32,56,44,10,32,32,32,99,111,108,111,114,58,32,39,98,108,97,99,107,39,44,10,32,32,32,112,97,100,100,105,110,103,82,105,103,104,116,58,32,51,48,44,10,32,32,32,98,97,99,107,103,114,111,117,110,100,67,111,108,111,114,58,32,39,35,102,48,102,48,102,48,39,44,10,32,32,32,104,101,105,103,104,116,58,32,52,53,44,10,32,32,32,109,97,114,103,105,110,84,111,112,58,32,45,53,44,10,32,32,32,119,105,100,116,104,58,32,49,50,53,44,10,32,125,44,10,32,105,99,111,110,67,111,110,116,97,105,110,101,114,58,32,123,10,32,32,32,116,111,112,58,32,39,53,48,37,39,44,10,32,32,32,114,105,103,104,116,58,32,48,44,10,32,32,32,116,114,97,110,115,102,111,114,109,58,32,91,123,32,116,114,97,110,115,108,97,116,101,89,58,32,45,49,48,32,125,93,44,32,47,47,32,65,100,106,117,115,116,32,116,104,105,115,32,118,97,108,117,101,32,97,115,32,110,101,101,100,101,100,32,116,111,32,99,101,110,116,101,114,32,116,104,101,32,105,99,111,110,10,32,125,44,10,32,98,117,116,116,111,110,58,32,123,10,32,32,97,108,105,103,110,73,116,101,109,115,58,32,39,99,101,110,116,101,114,39,44,10,32,32,106,117,115,116,105,102,121,67,111,110,116,101,110,116,58,32,39,99,101,110,116,101,114,39,44,10,32,32,112,97,100,100,105,110,103,86,101,114,116,105,99,97,108,58,32,49,50,44,10,32,32,112,97,100,100,105,110,103,72,111,114,105,122,111,110,116,97,108,58,32,51,50,44,10,32,32,98,111,114,100,101,114,82,97,100,105,117,115,58,32,52,44,10,32,32,101,108,101,118,97,116,105,111,110,58,32,51,44,10,32,32,98,97,99,107,103,114,111,117,110,100,67,111,108,111,114,58,32,39,98,108,97,99,107,39,44,10,125,44,10,116,101,120,116,58,32,123,10,32,32,102,111,110,116,83,105,122,101,58,32,49,54,44,10,32,32,108,105,110,101,72,101,105,103,104,116,58,32,50,49,44,10,32,32,102,111,110,116,87,101,105,103,104,116,58,32,39,98,111,108,100,39,44,10,32,32,108,101,116,116,101,114,83,112,97,99,105,110,103,58,32,48,46,50,53,44,10,32,32,99,111,108,111,114,58,32,39,119,104,105,116,101,39,44,10,125,44,10,125,41,59,10,10,10,101,120,112,111,114,116,32,100,101,102,97,117,108,116,32,67,114,101,97,116,101,80,111,115,116,59,10,10,10,10
